Lecturer Iterator
Followed the class diagram roughly to implement the following:
IStudentIterator
IStudentList
StudentIterator
StudentList
classes

Currently, inputting a staff ID on login will iterate through the list of lecturers. If it matches, you will be logged in and welcomed with your lecturer name

Functionality of Lecuter Iterator will remain basic until further progress in project is made and changes are needed.

Changed some spacing in StudentList.cs for clarity

Changed Lecturer.cs to a public class and changed spacing

Made changes in program.cs to add lecturers to a LecturerList lecturerCollection.

Then made changes in the login code to use the iterator to check if there is a match in the lecturerCollection.

Also changed names of the studentiterator instance in program.cs for clarity